Renaming executor files to executor.py
