+++ parsl/executors/__init__.py
@@ -1,7 +1,7 @@
 from parsl.executors.threads import ThreadPoolExecutor
 from parsl.executors.ipp import IPyParallelExecutor
-from parsl.executors.high_throughput.htex import HighThroughputExecutor
-from parsl.executors.extreme_scale.exex import ExtremeScaleExecutor
+from parsl.executors.high_throughput.executor import HighThroughputExecutor
+from parsl.executors.extreme_scale.executor import ExtremeScaleExecutor
 
 __all__ = ['IPyParallelExecutor',
            'ThreadPoolExecutor',
